Uat

See merge request riseplus/ca-data-center/ca-data-center-api!1

+++ lib/tasks/reset_data.rake
@@ -2,7 +2,7 @@ namespace :db do
   desc "Reset primary key sequences and delete all data from all tables"
   task reset_data: :environment do
     ActiveRecord::Base.connection.tables.each do |table|
-      next if table == 'schema_migrations' || table == 'ar_internal_metadata'
+      next if table == 'schema_migrations' || table == 'ar_internal_metadata' || table == 'provinces' || table == 'districts' || table == 'sub_districts' || table == 'post_codes' || table == 'address_codes' || table == 'cancer_form_statuses'  || table == 'icdos' || table == 'hospitals' || table == 'topography_codes' || table == 'address_codes' || table == 'marital_statuses'
 
       # ลบข้อมูลในตาราง
       ActiveRecord::Base.connection.execute("TRUNCATE TABLE #{table} RESTART IDENTITY CASCADE")
